services/router/requirements.txt
anyio==4.13.0
click==8.3.3
fastapi==0.100.0
h11==0.16.0
idna==3.13
kafka-python-ng==2.2.3
prometheus_client==0.21.0
psycopg2-binary==2.9.12
pydantic==1.10.13
python-dotenv==1.0.1
redis==5.0.1
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.30.0
